Move groupTypes function to ggrc utils from mappings
