+++ src/ggrc-client/js/components/object-generator/tests/object-generator_spec.js
@@ -6,7 +6,7 @@
 import RefreshQueue from '../../../models/refresh_queue';
 import Component from '../object-generator';
 import Program from '../../../models/business-models/program';
-import * as Mappings from '../../../models/mappers/mappings';
+import * as modelsUtils from '../../../plugins/utils/models-utils';
 
 describe('object-generator component', function () {
   'use strict';
@@ -74,14 +74,14 @@ describe('object-generator component', function () {
         });
 
         it('returns grouped snapshotable objects', () => {
-          spyOn(Mappings, 'groupTypes')
+          spyOn(modelsUtils, 'groupTypes')
             .and.returnValue('grouped snapshotable objects');
 
           expect(viewModel.availableTypes())
             .toEqual('grouped snapshotable objects');
-          expect(Mappings.groupTypes)
+          expect(modelsUtils.groupTypes)
             .toHaveBeenCalledWith(GGRC.config.snapshotable_objects);
-          expect(Mappings.groupTypes).toHaveBeenCalledTimes(1);
+          expect(modelsUtils.groupTypes).toHaveBeenCalledTimes(1);
         });
       });
     });
