+++ src/ggrc-client/js/plugins/tests/ggrc_utils_spec.js
@@ -148,3 +148,4 @@ describe('getTruncatedList() util', () => {
     expect(result).toBe(expected);
   });
 });
+
